Updated Makefiles to work on Ubuntu 11.10 et al.
Apparently the order of arguments to the linker matters, especially now
that Ubuntu has the --as-needed flag turned on by default. Without this
fix, the linker would silently drop BLAS and LAPACK from the list of
libraries to link to because none of the files that preceded them on the
command-line require these libraries. Then the linker would die complaining
of missing symbols from the BLAS and LAPACK libraries it had just dropped!
This prevented FAME, GATPFit, and DASPK from compiling.

In the new Makefiles, the libraries to link to are now specified in a LIBS
variable that is placed at the end of the linker command, while other
arguments remain in the LDFLAGS variable at the beginning of the linker
command. I've also removed some requests to link to the pthread library,
since we don't do any threading in our Fortran codes.

Closes #224.