Don't apply k(T,P) high-P-limit check to isomerization reactions.
k(T,P) values are always a combination of direct and well-skipping effects.
(If there is no direct pathway then they are entirely composed of
well-skipping effects.) In the limit of infinite pressure the direct
paths should dominate, while the well-skipping paths become negligible.
However, the lowest T and highest P condition we use for testing may not
always be close enough to the high-pressure limit for this to be true.

An example of this is a three-well isomerization network with path
reactions A <=> B, B <=> C, and A <=> C. If the A <=> C barrier is very
high, the direct path becomes very slow, and the well-skipping path
A <=> B* <=> C dominates. The result is that the k(T,P) values are much
faster than the high-P limit of the direct path, since the well-skipping
route is itself much faster than the direct route. This is true even at
the lowest T and highest P FAME considers. Therefore, we should not
consider isomerization reactions when performing this check.

This should take care of the high-P limit "warnings" with the largest
discrepancy. Many actual warnings still remain, and these are in fact
caused by to numerical roundoff as a result of the energy discretization,
so decreasing the grain size will get rid of most of these. (Incidentally,
in RMG-Py I've implemented a scheme for using different energy grains at
different temperatures, which seems to further reduce the frequency of
these warning messages.)