Must now specify both grain size and number of grains in FAME input files.
This feature allows finer control over the energy grains by specifying
both a minimum number of grains and a maximum grain size. FAME will use
whichever of these gives a more accurate result (by causing more grains of
a smaller size to be used).

Also updated the Java code to write FAME input files in the new syntax.
Note however that we don't currently provide options for setting the
grain size in RMG jobs. The default value of 251 should be a good choice
for most jobs.