Use simpler algorithm for automatically choosing the maximum energy grain.
Recent experience has suggested that the complicated equilibrium
distribution computation used before is overkill, and that simply choosing
a certain number of kB*T above the maximum ground-state or transition-state
energy is sufficient. Furthermore, if we intend to use different energy
grains at each temperature, using the old method will be too slow. I've
set the cutoff to 40*kB*T, which seems to be enough for almost every
network I've tried (i.e. the k(T,P) values are no longer sensitive to the
choice of the multiplier).