Compute densities of states using new energy grains.
This is the first step toward choosing different energy grains at each
temperature (to spend our computational time more efficiently). The idea
is to compute the densities of states once using a set of energy grains
that we can map to the grains used at each temperature. For this we choose
grains that span the energy range at the maximum temperature but with a
grain size equivalent to that of the minimum temperature. We also leave
the computed densities of states in the array such that the first value
is always at the E0 of the isomer, leaving the shifting to a common zero
of energy to a later step.

Note that, since we will now be using different energy grains at each
temperature, we must pass the desired grain size and number of grains to
network_calculateRateCoefficients().