Added method for mapping densities of states at each temperature.
This is the next step toward using different energy grains at each
temperature. The grains do not need to align between those used for the
densities of states and those used at each energy; when they don't, the
mapping algorithm is smart enough to interpolate as needed.

Since the energy grains are now changing at each temperature, we need to
reallocate the arrays of quantities that depend on energy: densities of
states, equilibrium distributions, and microcanonical rate coefficients.