Compute densities of states for all configurations in FAME.
Previously I believed that the density of states calculations was expensive
enough that we wanted to minimize the number of times that we do it.
However, nowadays I feel that this is not so expensive (especially in
Fortran), and that it's more important to spend a little more time to be
more accurate. By computing the densities of states for each isomer,
reactant channel, and product channel, we can now compute the k(E) using
ILT in the direction the kinetics are specified, which eliminates one
source of numerical discretization error. (This feature coming separately.)