Compute k(E) in direction kinetics is given in FAME.
Now that we have the densities of states for all isomers, reactant
channels, and product channels, this is something we can do. Previously we
would have to get the reverse k(T) before applying the ILT algorithm, which
introduced an unnecessary source of numerical discretization error, and
caused the k(T,P) values to deviate from the high-P limits at high
pressure. The new approach -- along with the better ILT algorithm for n > 0
and the use of different energy grains at different temperatures -- causes
the discretization error to decrease significantly, so that a relatively
low number of energy grains can continue to be used.