Save a Chemkin file in PopulateReactions.
One of the reasons to run PopulateReactions is to generate some reactions
by hand to add to a kinetics model. This process would be much easier (at
least with the workflow the group currently uses) if we had a Chemkin file
produced as a result of running PopulateReactions. This commit adds such
functionality, although it's a bit hackish, since we don't appear to
formally create a reaction model or reaction system, which is what the
Chemkin file writer requires.