Undo attempt to interpolate while applying ILT method for n /= 0.
The implementation was found to give NaN in some situations for the first
element of k(E). Since we are now doing adaptive grain sizing, I don't
think the small increase in accuracy is worth the trouble, especially given
the approximate nature of the methods used to compute k(T,P). (It should
be noted, however, that the formula does seem to work for the system in
question in RMG-Py.)

Closes #229.