Added specfun.f90 to FAME for gamma function.
gamma is only an intrinsic function in gfortran version 4.3 or later. This
commit provides a set of special mathematical functions, including the
gamma function, so that FAME will still build with older compilers.

Note: specfun is licensed under the LGPL which we should note in the
documentation. Website for specfun:

http://people.sc.fsu.edu/~jburkardt/f_src/specfun/specfun.html

Closes #240.