also check aromatic species for equivalence in equalsAsChemGraph even if one has resonance isomers and one doesn't
(hopefully) closes #239 ;
we check here whether the molecule is aromatic because currently in RMG, it seems that some representations of a given aromatic radicals may be considered to have resonance isomers while other representations do not (probably depending on how the species/chemgraph was generated (e.g. whether generated from a form already known to be aromatic with B bonds vs. a newly formed aromatic species without pre-existing B bonds ));
this is admittedly somewhat of a hack, but there is no obvious easy alternative and it seems that there is a large block of code here (ReactionTemplate.calculateForwardRateConstant(), Structure.equalsAsChemGraph(), and this function Structure.cgPresentInList() ) that is only called by generateReverseForBackwardReactions for intraH_migration reactions