Better way of choosing direction to return k(T,P) values from FAME.
The previous approach tried to choose the direction with the larger
k(T,P) values. However, this might differ depending on the temperature,
and is certainly not simple to predict a priori. The new method uses the
ground-state energies of the reactants and products to decide, which is
much more straightforward.