Don't check T and P range if not defined.
The TMIN, TMAX, PMIN and PMAX are static variables belonging to the entire
class of PDepRateConstants, and are only defined if running with PDep on.
If running with PDep off, there are no validity limits, and yet we still
may want to evaluate PDepRateConstants coming from a seed mechanism or PRL.
In this case, we just skip (or always pass) the validity test.

Closes #242 and #253