Make GATPFit compile with gfortran.
I reverted one of the changes in the previous commit. I like the change, it looks nice,
but I think it uses something in the Fortran 2003 standard, and this file is in F77 format,
and I couldn't for the life of me get it to compile in gfortran. This compiles, and for me, works.