Small tweak to the radical group tree for allylic ketones.
Previously, if you had a tertiary radical allyl to a ketone carbonyl, the
estimated thermo would only account for the tertiary substitution, and
not the resonance stabilization the carbonyl provides. We did already
have a group for an aldehyde carbonyl. This commit uses the aldehyde
value for both ketones and aldehydes, and that's almost certainly better
than just forgetting about the C=O contribution entirely.