RMG calls the new looping Frankie once and passes data back and forth.
The new frankie.exe runs continuously, taking in data from STDIN and 
returning results on STDOUT. This commit updates the FrequencyGroups.java
file to interact with it accordingly.

I tried to maintain all the existing functionality, which means that 
the check for convergence is done by reading the frankie.log file from 
disk. This is unfortunate, and maybe should be fixed next.