QM for fused cyclics only
Since fused cyclics are the class of molecules that are predicted worst
by Benson extrapolation, I introduced a new flag for these class of
compounds.

Like the flag for cyclic compounds, the flag for fused cyclics searches
for fused ring atoms.

If the flag is activated, QMTP will only be used for compounds with
fused ring atoms. For species such as cyclopentane, the regular Benson
GA method will thus be applied.

The addition of this flag is both done in RMG as in TDE.