QMTP calls narrowed down to non aromatic fused ring systems
the flag qmOnFusedCyclicsOnly is changed to qmOnNonAromaticFusedCyclicsOnly
since it is believed that fused polycyclic aromatic rings such as
indene, are better predicted by Benson group additivity than QMTP methods
like PM3, MM4.

the philosophy behind is that naphtheno aromatic polycyclics have less
ring strain energy due to the planar structure of the aromatic rings.

A new method in Graph "containsAromaticRing()" searches the SSSRs
and returns true if at least one aromatic ring is found.

If this is the case, then QMTP is replaced by GATP as the thermo estimator
driver.

positively tested on 4-methylindene, fulvene, and naphthalene.