polycyclic ring strain corrections (PCRC) implemented
PCRC are part of Benson's GA framework and are superimposed on the
GAVs perceived in the molecule.

PCRCs are added as a new type of Thermo Group libraries. This implies that
ThermoGAGroupLibrary now has entries (Dictionary, Tree, Library) for PCRCs

The framework is copied from other types of groups, and adopts reader
methods used by the other groups too (eg. GAVs, NNIs, RSCs, HBI, etc...)

The interaction between the GATP and QMTP method is as follows:
-If a molecule is acyclic always GATP
-If a molecule is cyclic but contains no fused atoms, then separate
Ring strain corrections can be applied additively. (eg. cyclopentyl-cyclo-
pentene). The GATP method is therefore exclusively used for non fused
ring systems. GATP iterates over the RSC library, not the PCRC library.

-If a molecules contains fused ring atoms, then the GATP iterates over the
PCRC library, not the RSC library since separate monocyclic RSCs will
probably not be as adequate as PCRCs. If a PCRC is identified that
corresponds to the system in the current molecule, this PCRC is added.
If none of the PCRC entries is matched, then QMTP method is called.

The QMTP flag is still active, meaning that the user can choose to never
call QMTP at all.