Always add Ar, N2, Ne, He to species list in generated Chemkin files.
This is because these sometimes occur as colliders in third-body,
Lindemann, and/or Troe kinetics. Oddly, we were always adding the thermo
of these colliders, but not always adding them to the species list.

This is a hack, of course, but it's at least consistent with what was
already there. :P

Closes #264.