only use polycyclic ring corrections if complete ring system is matched.
prior to this patch, a matched polycyclic ring system in a molecule
would be applied, even if the molecule contained more ring atoms
than the number of nodes in the hit functional group structure.

after this patch, in case of a matched polycyclic ring system,
the total number of ring atoms in the molecule must match the
total number of atoms in the matched polycyclic ring system,
assuring that even larger polycyclic ring systems in the molecule
do not exist.