prevent the "trickle-up" system for single ring corrections
as with the other group contribution trees in RMG, a ring correction
would be applied for the most specific ring structure specified in the
tree. In case there are no L3 nodes that match the ring system, then
the same principle would be applied to L2 nodes, all the way up to L0 nodes
if only the most generic node could be matched.

This behaviour has the consequence that string ring systems like
C1=C=CCC1 with 2 double bonds in the ring, will hit the "FiveMember"
node because none of the other nodes match. Since the "FiveMember" node
refers to the cyclopentane ring correction values, this strange ring gets
actually pretty normal ring correction. hence the consequences for model
generation with faulty thermochemistry assigned for specific species.

This trickle up system is halted by checking if the matched ring node
has a level 1 or higher. If so, then none of the specific nodes could be
hit and we should better use the QMTP instead of Benson.