set the useQM to true, always.
due to a previous commit that forces the thermo data estimation to fall
back to QMTP if only a non-trivial (e.g. Six-Member) ring correction was
matched, the QMTP module should always be active, even when the useQM flag was
was set to false.

I therefore set the useQM flag temporarily to "true", hard-coded.

Morever, it does not make sense to NOT have the QMTP activated when
dealing with cyclic species.

I don't see this to cause too many problems because when models with
only acyclic species are generated the QMTP class will not be called
because the ChemGraph.isAcyclic() will redirect to the GATP class.