Added MaxAtomsForPressureDependence option to condition file.
Once implemented, this option will turn off consideration of pressure
dependence for molecules with more than the specified number of atoms.
This has the potential to make certain jobs go much faster and use much
less RAM in the process. The documentation and example condition file
with all options have been updated accordingly.

If the option is not given, the previous behavior (run pdep for
everything) remains the default.