isReactionPressureDependent() now also uses the maxAtoms criterion.
The FastMasterEquation.maxAtoms attribute stores the total number of
atoms above which pressure dependence is assumed to be negligible, as
set from the condition file. (It is initialized to a very large number
so that, by default, all systems are considered as pressure dependent.)
The FastMasterEqn.isReactionPressureDependent() method now honors this
value when evaluating the pressure dependence of a given reaction.

A helper method for determining the total number of atoms involved in a
reaction was also added as Reaction.getAtomNumber(). (Not my preferred
name, but consistent within the naming scheme currently used by
RMG-Java.)