removed bug: identical ring strain corrections
the bug would not count multiple identical ring strain corrections because
it did not look up the value of the key of the map "deepestStackMap"
that contain the key-value pair for every ring correction.

so you we perceived multiple identical ring strain corrections, but we
did not APPLY them.

this is now fixed.