Python scripts add things to front of sys.path instead of end.
It is important that the modified version of cclib that is bundled with RMG
is used, in preference of an official or external version that may exist
on your python path. I assume this may be true of the MOPAC scripts etc too,
so have changed all @gmagoon's `sys.path.append(foo)` calls into 
`sys.path.insert(1, foo)` calls to make sure things are at the front of the path.