Updated H_Abstraction tree for X_H_or_Xrad_H
Union of X_H and Xrad_H was added to the dictionary in 296a70b236554b6824b8556b72f0bbdb74a36256, but
the tree was never updated appropriately. This is causing problems
on importing to Py, as the rate rule X_H_or_Xrad_H;Y_rad_birad is
brought in without a matching group in the tree.