TDGenerator and implementations added
TDGenerator is an abstract class that serves as the template class for
all classes that implement some kind of strategy to estimate thermochem-
istry of molecules.

3 implementations are added; all of them implement the concrete method
"generateThermo(ChemGraph)":

1) BensonTDGenerator: use only Benson GA for TD. if adequate RSCs, and
polycyclic RSCs are not found, then you print an error message to the logger
but it doesn't crash. It will use the "less accurate" estimations.

2) QMForCyclicsGenerator: uses QM methods for cyclic molecules. If QM
attempts repeatedly fail, then it falls back to the BensonTDGenerator.
Acyclic molecules are always estimated through BensonGA.

3) HybridTDGenerator: uses Benson GA to estimate properties. If adequate
(poly)RSCs could not be found, it falls back to the QM Generator. If the
latter fails, it falls back to the good ol' BensonGAGenerator.