ChemGraph contains new Thermo estimation scheme
The new scheme encapsulates the different thermo estimation strategies.
Old QMFlags attributes are replaced by a single new one, called
TDSTRATEGY. This String can have three different values:
-BensonOnly
-QMForCyclics
-Hybrid

and corresponds to the different implementations of TDGenerator.

Moreover, in GATP a class attribute "Map<ThermoGAValue, Integer> monoCyclicRSCs"
is created that is filled with data when the method "getRingCorrections"
is called.

ReactionModelGenerator is modified to read the new line and not read
the older deprecated lines.

In ReactionModelGenerator InChi's are only written when the QMForCyclics
or Hybrid strategy are chosen.