Updated thermo of iC4H7OJ(51) in CBS_QB3_1dHR thermo library.
We found that the adjacency list was not for the species we thought it
was for. I re-ran the QC calculations for the actual species and saw a
small but significant change in the H298 and S298.

Note: This species happens to be rather important in the DIPK mechanism,
which is how I found the issue.