Some tweaks to the PDepNetwork.addReactionToNetworks() method.
The tweaks help ensure that the reaction is added to the proper network
in the direction the kinetics are defined in, which is preferred by
FAME from a numerical standpoint, as this direction tends to have the
larger value of k(T). Also don't forget to set the redundancy when
creating the PDepReaction equivalent of the Reaction object.