Added lines for printing averaged rate rules created at start of job.
This has been useful for debugging (e.g. when reimplementing the rate
rule method for kinetics estimation in RMG-Py). I've commented it out
since it adds a lot of distracting information to the start of the job,
but I wanted to commit it in case more debugging is required in the
future.