Revert commit a505cdae7cefcdfda40a9bc3f753745fadde4551
It appears that occasionally a reaction with undeterminable kinetics
reaches the ReactionTemplate.findClosestRateConstant() method. Before
commit a505cdae, a RateConstantNotFoundException was thrown when this
occurred, which was handled by (silently) dropping the reaction from
further consideration. In a505cdae this behavior was "temporarily"
changed to return an Arrhenius expression with all zeros, for reasons
since forgotten. This is problematic with pdep on, as FAME expects
physical kinetics to be returned.

As far as I can tell, reverting this commit does no harm while fixing
the above issue.