Added hardcoding of Substitution_O reaction family.
Since Substitution_O is its own reverse, the atom labels of *2 and *3
need to be swapped after applying the reaction recipe, or else the
reverse kinetics will not be estimatable. This will manifest as a large
number of errors of the form "ERROR: Couldn't find the rate constant for
reaction:" in the log. (Naturally, RMG will just merrily continue even
after hitting this supposed "error".)

Closes #276.