Made HO2_Elimination_from_PeroxyRadical family reversible.
I could not find any reason for why the family was made irreversible in
the first place in several Green group member theses from the last
several years. (In fact, the brief section in Catherina Wijaya's thesis
talks about the reverse reaction a bit.) Looking at the git history, it
appears this family has been irreversible since it was first added to
version control in 2005.