Added BensonRingCorrections class for storing metadata about the corrections.
This class currently stores the set of Benson-like ring corrections, along
with a flag for indicating if the corrections are approximate. This allows
us to both apply the best available ring correction (BensonTDGenerator)
while still falling back to QMTP if the corrections are approximate
(HybridTDGenerator). In practice, it allows us to return more than just
the corrections themselves from ThermoGAGroupLibrary.findRingCorrections().