Writing condition file for restarting a RMG job
RMG now also writes an additional file in Restart folder
which can be used as a condition file for restarting a
job. This is done as current RMG-Java restart feature
is hard to maintain and breaks with every change to main code.

This should also be faster than reading all the restart files
from disk.

Please enter the commit message for your changes. Lines starting