Run InChI in the RMG_JOB_SCRATCH/InChI directory.
Default behaviour should be the same as before, because
RMG_JOB_SCRATCH defaults to $PWD if it is not defined.

However, if you DO set RMG_JOB_SCRATCH then you can put it
on a faster drive, and keep your temporary InChI files out
of the way of your results.