Adding rough rate rules for C(R1)(R2)=O + ROO
Earlier RMG is using rules for RO for adding to double
bond. Josh and Shamel think this must be much slower (different)
for ROO species. For now using a very rough estimate for these
reactions based on JWA calculation for Acetone + HO2.
This must be updated with better rules in the future.