QMLibrary files are now opened, written, and closed for each species.
This should allow multiple concurrent runs of RMG to write to the same
QM thermo library. (They won't read from each other's results until they are 
restarted, but at least their results will all be saved in one place).
This could lead to the same species showing up several times in the list, but 
this shouldn't be a problem.