RestartConditionFile now saved in RMG_JOB_OUTPUT directory.
Instead of the Restart directory. It is also now saved even if you 
request not to save Restart. I decided this is reasonable because it's very 
cheap to write this file (Writing the other Restart takes a long time, which 
is why people turn it off).

Also made a few tweaks. and stripped some whitespace.