Better checking of ring strain corrections involving thermo libraries.
It turns out that the GATP class also checks the thermo libraries for the
species thermo before reverting to group additivity. If the species is found
in a library, it should not have any monoCyclicRSCs, which would lead to a
NullPointerException since some subsequent code assumed that monoCyclicRSCs
is always non-null. To fix this, I added a check for whether the thermo is
coming from a thermo library, which does not perform the monoCyclicRSCs
check if so.

I also added a resetting of the monoCyclicRSCs attribute at the start of
generateThermoData() to make sure the value is cleared from molecule to
molecule. (I think this was inadvertently hiding the error.)

Closes #281.