Updated OCOO rule in HO2_Elimination_from_PeroxyRadical family.
The old rule was an estimate made by MRH made by generalization of a rule
for CH3CHO + HO2 = HOCH[OO]CH3. The new rule is from a paper by Hermans et
al. (2005), and is for CH3COCH3 + HO2 = CH3C(OH)(OO)CH3. The two rules are
actually within an order of magnitude or so anyway, so there's not a big
change here.