Updating the restart condition file
Earlier the restart condition file would write in all
the species in the core. This would lead to unintentional
consequence of all species specified in the seed
reacting with each other and generating all possible reactions
(i.e restarting with GenerateReactions = on for all seeds).
The restart condition file now skips over the species specified
in the seed as they will be added in when the model is
initiated again in RMG.