Remove references to FileNotFoundException from ReactionModelGenerator.
Every occurrence was in a catch block that was immediately followed by another
catch block for IOException, with both blocks containing exactly the same code
(usually Logger.logStackTrace()). Since FileNotFoundException is a child of
IOException, this code is redundant. (A few instances were causing warnings to
appear during compilation, which motivated this change.)