Some restructuring of the Cyclic_ether_formation tree.
The new tree makes R2OOH, R3OOH, etc. and R2OOR, R3OOR, etc. mutual siblings.
The motivation for this is that the previous definition of the tree made
when trying to allow for OR leaving groups in cyclic ether formation caused
a change in how the kinetics were estimated for the original OH leaving
groups. The particular reaction I noticed this for was

C(COO)(C)C(=O)[C](C)C -> C(CO1)(C)C(=O)C1(C)C + [OH]

For the above reaction the new kinetics averaging scheme is almost certainly
worse, as it now averages over R2OOH, R3OOH, and R4OOH ring sizes rather
than over the other functional group.

Probably we need to look more at why the previous change to the tree changed
the way the kinetics are estimated as well, but that is probably a
significant undertaking.