In ThermoDataEstimator, don't require other QM lines when using BensonOnly.
When using the BensonOnly option for estimating thermo, the other lines are
meaningless. After this commit, you should not specify those lines if using
the BensonOnly method. Note that, in some of the other main scripts, all of
these lines are optional, but in the current ThermoDataEstimator, you must
at least specify BensonOnly.