Only check for RDBASE environment variable when it is needed.
Previously this variable would be checked when the QMTP class was
instantiated. Since it follows the singleton pattern, this instantiation
always occurs, even for BensonOnly jobs. Since we don't need RDBASE in
this case, it is better to perform the check at the point of use of that
environment variable instead.

This also eliminates the appearance of the RDBASE warning when running
regular RMG jobs using BensonOnly for thermo estimation.

Closes #282.