Atom types are now case sensitive.
This must be done in order to allow both the Cs (carbon with only single
bonds) and CS (carbon with double bond to sulfur) atom types to coexist.
Without this, using group additivity or rate rules to estimate parameters
could lead to incorrect leaves in the tree due to the mistaking of CS for
Cs.

Closes #284.