Print warning instead of crashing weirdly if trying to read from Tree before it's loaded.
It used to give a bizarre "TreeNode can't be cast into HierarchyTreeNode" type of crash,
because the blank HierarchyTree would have a root element that's a TreeNode (from the parent
constructor). Perhaps this should be an exception not a warning, but at least it now tells
you the problem which may help debugging.