Updates to the Abraham group estimation groups database from ajalan/autoxidation_rebased
This is all Amrit's (@ajalan) updates to the Abraham groups Dictionary/Library/Tree
currently on his autoxidation_rebased branch. This should bring the master up to date
in this regard. (Although there are other changes on that branch that should also be merged).
I believe some of the changes fix 1,2-diols. I'm not sure about the others.