Updated DASSL to handle up to four products in reactions.
The reaction arrays for regular, third body, Lindemann, and Troe reactions
needed to be updated to include another element for the new product. (Aside:
why aren't these arrays two-dimensional?)

DASPK has not yet been updated, since it would be even more involved, and
I'm not set up to test it at the moment. However, the sorts of changes made
here should be similar to changes that would be needed there.

Note that the maximum number of reactants and products is still hardcoded in
many places.