Clossing aromaticity perception for generation of reactions
To allow species like cyclohexadienyl radical to go to benzene
we need the aromatic species to be represented by single-double
bonds.

However to get the correct thermo and transport properties we still
should convert it back to Cb representation. Currently we will just make
a copy of the graph and percive aromaticity on it to get the correct
thermo and transport properties.

Done with help of Aaron Vandeputte