Earlier code would break for fused aromatic ring with radical
If one of the resonance isomers has fused rings with
one ring classified as aromatic and the second as
non aromatic then a shared C can have a B bond and
a D bond C {*,B} {*,B} {*,D}
causing the valency check to fail.

To avoid this we now have additional checks
which convert the other ring to be classified as non-aromatic.

This is for academic interest only :) .... since Aaron thinks
the most stable isomers will have both rings classified as
aromatic.