Aarons update for aromaticity perception in fused rings
Earlier bug fixes did not classify anthracene as aromatic anymore.
The algorithim is now modified such that the second aromaticity check
updates a new list which is sent for reclassification by Sandeep's
aromaticity algorithim. This ensures that when one ring classification
is changed from aromatic to non-aromatic we recheck all the other rings
to see if their classification has changed. This a recurssive and
converges when the aromatic List and updated aromatic List match.
Only then we convert the aromatic species to B bonds.

This change was almost entirely written by Aaron Vandeputte .. I have
just made it more streamlined and cleaned up the code a bit.