Fixing bug for calculating flux when the reaction is given in a reaction library
For bimolecular reactions coming from reaction libraries
RMG did not check the direction in which the reaction was given.
This led to incorrect flux being calculated.

In Reaction library B + C = A  k1
RMG finds reaction A = B + C
RMG old flux = k1 * [A]
Now it checks the direction and calculates the flux
RMG new flux = k1* Keq * [A]

We havent checked what happens in a pdep reaction!

It must also be noted that if in a reaction library,
duplicates reactions are given in different directions
then this fix will fail, since we only have one boolean
to check for every reaction instance !!