Turning "on" family Intra Diels Alder 2
This family is important for pyrolysis chemistry, large
alkane chains with multiple unsaturated sites can add
on the double bond to lead to rings eventually leading
to aromatics.

Note: Limited testing has shown this family not to break
current RMG usuage. If it breaks normal usuage for users
the default setting should be turned off