Changing head node value for Diels Alder Family
Most of the calculated rates have ~ 20-25 kcal/mol barrier,
SSM thinks the head node is too fast current as it set
barrier less ~ 5E+9. The rank of 0 is a special case
where the family is not averaged and every estimate defaults
to the head node (changing rank to 5 might change this back to 0).