Fixing bug of reading in thermo for aromatic species in primary thermo library
Since we converted the thermo_graph to B bonds it would not match the
chemgraph in the thermo library. The current implementation according to me
works this way

1) Create copy of chem_graph and store it in thermo_graph.
2) Check whether thermo_graph is aromatic and convert to B bonds.
3) If thermo_graph is aromatic chem_graph is also aromatic, set the flag.
4) Generate thermo for chem_graph, if this is coming from primary thermo
library then we are set, no need to generate thermo from the B bonded
thermo_graph.
5) If chem_graph is not coming from primary thermo library and chem_graph is
aromatic we need to generate the thermo on the B bonded thermo_graph

When chem_graph is not aromatic and has multiple resonance isomer
we need to check if thermo_graph is from primary library and also update the
attribute for the chem_graph.

Tested for Toluene, Fulvene