More helpful message for Pdep failure
My jobs suddenly started failing while running pdep jobs,
this was because one of the libraries I was using had
irreversible reaction. This commit will give a more helpful
message when it fails.

Irreversible exit channel is typically allowed in pdep codes
like variflex, I am not sure if we should just check for
whether reaction is irreversible and only then try to reverse it.
Or make an irreversible reaction also reversible in order to
maintain thermodynamic consistency (?).