Fixing missing thermo comments
Because we made a copy of the chemgraph to get the thermo and
were calling the comments from this new graph, there were cases - when the
molecule was not aromatic it would return an empty string.

We now only call the comments from the thermo_graph if the species is aromatic and not present in a thermo library.

Sigh another Sunday (10:00 pm) bug fix!