Adding the option to set max C atoms for a Pdep Network
This option provides more flexibility to determine if we need
to treat a reaction as pressure dependent. Ideally we should not
recommend using this option since whether to treat a reaction as
Pdep or not depends on the total number of atoms in a reaction (i.e net degree
of freedome determines whether a molecule can undergo internal relaxation).

However for some practical consideration when your fuel is unsaturated
and leads to a lot of aromatic species (mostly heavy atoms due to decrease
in saturation) we might want to specify the Carbon number or total number of
Heavy atoms (not yet implemented).